move some files and run fmt
